+++ internal/timeutil/util.go
@@ -12,38 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package bufutil
+package timeutil
 
 import (
-	"crypto/rand"
-	"fmt"
-	"io"
 	"math"
-	"net"
-	"strconv"
-	"strings"
 	"time"
 )
-
-func IsValidIPAddress(addr string) bool {
-	return net.ParseIP(addr) != nil
-}
-
-func GetIPAndPort(addr string) (string, int32) {
-	var port int
-	var err error
-	parts := strings.Split(addr, ":")
-	if len(parts) == 2 {
-		port, err = strconv.Atoi(parts[1])
-		if err != nil {
-			port = 5701 // Default port
-		}
-	} else {
-		port = -1
-	}
-	addr = parts[0]
-	return addr, int32(port)
-}
 
 func GetTimeInMilliSeconds(duration time.Duration) int64 {
 	if duration == -1 {
@@ -69,16 +43,4 @@ func ConvertMillisToUnixTime(timeInMillis int64) time.Time {
 		return time.Unix(0, timeInMillis)
 	}
 	return time.Unix(0, timeInMillis*int64(time.Millisecond))
-}
-
-// NewUUID generates a random uuid according to RFC 4122
-func NewUUID() (string, error) {
-	uuid := make([]byte, 16)
-	n, err := io.ReadFull(rand.Reader, uuid)
-	if n != len(uuid) || err != nil {
-		return "", err
-	}
-	uuid[8] = uuid[8]&^0xc0 | 0x80
-	uuid[6] = uuid[6]&^0xf0 | 0x40
-	return fmt.Sprintf("%x-%x-%x-%x-%x", uuid[0:4], uuid[4:6], uuid[6:8], uuid[8:10], uuid[10:]), nil
 }
